\name{PredictAnswer}
\alias{PredictAnswer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Answer prediction
}
\description{
Predicts answer.
}
\usage{
PredictAnswer()
}
%- maybe also 'usage' for other objects documented here.
\details{
More details.
}
\value{
Answer
}
\references{
None
}
\author{
Test
}
\note{
Test
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
Test
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
PredictAnswer()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ answer }
\keyword{ prediction }% __ONLY ONE__ keyword per line
